% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset}
\format{
A data frame with 9999 genetic associations
}
\usage{
data(toy_data)
}
\description{
A data frame with 9999 genetic associations
}
\details{
\itemize{
\item rsID. SNP name
\item freq. allele frequency of effect SNP
\item A1_beta. effect size
\item Analysis. adds colour to the type of analysis (e.g. GWAS, Sequencing)
\item Gene. Candidate gene name
\item N.
\item trait. ToyDataPheno
}
}
\keyword{datasets}
